+++ tests/fixtures/shells/env_setup.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+export FROM_ENV=yes
+export DZ_TEST_MARKER=setup_ran